+++ CHANGELOG.md
@@ -1,8 +1,29 @@
-# Changelog
+# Changelog 
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+
+## Unreleased
+
+## 4.56.0
+
+### Added
+- Add bank_name and external_account_id for DEBIN tracks
+- New payment method model support
+- Added Congrats Extra Info (IFPE regulations)
+
+### Fixed
+- Fixed discountToken in discounts for payment methods
+- Inclusão de flag para identificar se o pagamento já foi processado, evitando um possível pagamento duplicado.
+- Fixed CFTNA label for installments for MLA
+- Fixed deep_link key (tracks)
+
+### Changed
+- Hide duplicate value when payment method is debit card and splitConfiguration is null
+
+### Removed
+- Remove unnecessary tracks for path: /px checkout/dialog/dismiss
 
 ## v4.55.1